+++ cmd/utils/fetch_accounts_data.go
@@ -7,25 +7,6 @@ import (
 	"github.com/dymensionxyz/roller/config"
 )
 
-func GetHubRlyAccData(cfg config.RollappConfig) (*AccountData, error) {
-	HubRlyAddr, err := GetRelayerAddress(cfg.Home, cfg.HubData.ID)
-	if err != nil {
-		return nil, err
-	}
-	HubRlyBalance, err := QueryBalance(ChainQueryConfig{
-		RPC:    cfg.HubData.RPC_URL,
-		Denom:  consts.Denoms.Hub,
-		Binary: consts.Executables.Dymension,
-	}, HubRlyAddr)
-	if err != nil {
-		return nil, err
-	}
-	return &AccountData{
-		Address: HubRlyAddr,
-		Balance: HubRlyBalance,
-	}, nil
-}
-
 func GetSequencerData(cfg config.RollappConfig) ([]AccountData, error) {
 	sequencerAddress, err := GetAddressBinary(KeyConfig{
 		ID:  consts.KeysIds.HubSequencer,
