+++ relayer/query.go
@@ -1,8 +1,10 @@
-package sequencer
+package relayer
 
 import (
+	"github.com/dymensionxyz/roller/cmd/consts"
 	"github.com/dymensionxyz/roller/cmd/utils"
 	"github.com/dymensionxyz/roller/config"
+	"github.com/dymensionxyz/roller/sequencer"
 )
 
 func GetRolRlyAccData(cfg config.RollappConfig) (*utils.AccountData, error) {
@@ -10,7 +12,7 @@ func GetRolRlyAccData(cfg config.RollappConfig) (*utils.AccountData, error) {
 	if err != nil {
 		return nil, err
 	}
-	rollappRPCEndpoint, err := GetRPCEndpoint(cfg)
+	rollappRPCEndpoint, err := sequencer.GetRPCEndpoint(cfg)
 	if err != nil {
 		return nil, err
 	}
@@ -35,10 +37,29 @@ func GetRelayerAccountsData(cfg config.RollappConfig) ([]utils.AccountData, erro
 		return nil, err
 	}
 	data = append(data, *rollappRlyAcc)
-	hubRlyAcc, err := utils.GetHubRlyAccData(cfg)
+	hubRlyAcc, err := GetHubRlyAccData(cfg)
 	if err != nil {
 		return nil, err
 	}
 	data = append(data, *hubRlyAcc)
 	return data, nil
+}
+
+func GetHubRlyAccData(cfg config.RollappConfig) (*utils.AccountData, error) {
+	HubRlyAddr, err := utils.GetRelayerAddress(cfg.Home, cfg.HubData.ID)
+	if err != nil {
+		return nil, err
+	}
+	HubRlyBalance, err := utils.QueryBalance(utils.ChainQueryConfig{
+		RPC:    cfg.HubData.RPC_URL,
+		Denom:  consts.Denoms.Hub,
+		Binary: consts.Executables.Dymension,
+	}, HubRlyAddr)
+	if err != nil {
+		return nil, err
+	}
+	return &utils.AccountData{
+		Address: HubRlyAddr,
+		Balance: HubRlyBalance,
+	}, nil
 }
